requirements.txt
Flask
Flask_SQLAlchemy
Flask_Login
Flask_WTF
email_validator
pytest
pytest-cov
openai
gunicorn
python-dotenv
psycopg2-binary
OpenAI
PyPDF2
python-docx